requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
black==25.1.0
boto3==1.40.30
botocore==1.40.30
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cohere==5.18.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
distro==1.9.0
Django==5.0.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.1
django-ninja==1.4.3
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
eval_type_backport==0.2.2
executing==2.2.1
fastavro==1.12.0
fcm-django==2.3.1
filelock==3.19.1
firebase_admin==7.1.0
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.27
google-api-core==2.27.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
griffe==1.14.0
groq==0.31.1
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
invoke==2.2.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
logfire==4.7.0
logfire-api==4.7.0
markdown-it-py==4.0.0
mcp==1.14.0
mdurl==0.1.2
mistralai==1.9.10
msgpack==1.1.2
multidict==6.6.4
mypy_extensions==1.1.0
nexus-rpc==1.1.0
openai==1.107.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-ai==1.0.6
pydantic-ai-slim==1.0.6
pydantic-evals==1.0.6
pydantic-graph==1.0.6
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
swapper==1.4.0
temporalio==1.17.0
tenacity==9.1.2
tokenizers==0.22.0
tqdm==4.67.1
typer==0.20.0
types-protobuf==6.30.2.20250822
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.11.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
GitPython